requirements.txt
asgiref
Django
djangorestframework
pytz
sqlparse
tzdata
